% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaDatasetList.R
\name{beaDatasetList}
\alias{beaDatasetList}
\title{Retrieve List of Available Datasets from Bureau of Economic Analysis (BEA) API}
\usage{
beaDatasetList(ResultFormat = "json", beaKey = NULL)
}
\arguments{
\item{ResultFormat}{A character string specifying the desired format of the results.
Default is "json". Other possible values may include "xml" (check BEA API documentation for all options).}

\item{beaKey}{A character string containing the BEA API key. If NULL (default),
the function will attempt to retrieve the key using `getbeaKey()`.}
}
\value{
If ResultFormat is "json", returns a tibble (data frame) containing information about available datasets.
  For other formats, returns the raw API response.
}
\description{
This function retrieves a list of available datasets from the Bureau of Economic Analysis (BEA) API.
It requires a valid API key and can return results in different formats.
}
\examples{
\dontrun{
# Retrieve dataset list using default JSON format
datasets <- beaDatasetList()

# Retrieve dataset list using a specific API key
datasets <- beaDatasetList(beaKey = "YOUR-API-KEY-HERE")

# Retrieve dataset list in XML format
datasets_xml <- beaDatasetList(ResultFormat = "xml")
}

}
\references{
Bureau of Economic Analysis (BEA) API: \url{https://apps.bea.gov/API/docs/index.htm}
}
\seealso{
\url{https://apps.bea.gov/API/signup/} for BEA API registration
}
