% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaGDPbyIndustry.R
\name{beaGDPbyIndustry}
\alias{beaGDPbyIndustry}
\title{Retrieve GDP by Industry data from the Bureau of Economic Analysis (BEA) API}
\usage{
beaGDPbyIndustry(
  TableID = "",
  Industry = "",
  Frequency = "",
  Year = "",
  ResultFormat = "json",
  beaKey = NULL
)
}
\arguments{
\item{TableID}{Character string specifying the desired table ID.}

\item{Industry}{Character string specifying the industry code.}

\item{Frequency}{Character string specifying the frequency of the data (e.g., "A" for annual, "Q" for quarterly).}

\item{Year}{Character string specifying the year(s) for which data is requested.}

\item{ResultFormat}{Character string specifying the format of the results (default is "json").}

\item{beaKey}{Character string containing the BEA API key. If NULL, the function will attempt to retrieve it using `getbeaKey()`.}
}
\value{
A data frame containing the requested GDP by Industry data. If an error occurs, it returns an error message.
}
\description{
This function fetches GDP by Industry data from the BEA API based on the specified parameters.
}
\details{
This function requires a valid BEA API key. If not provided, it attempts to retrieve the key using `getbeaKey()`.
The function will display a warning and return an error message if the API key is invalid.
}
\note{
The function prints the statistic description and any notes returned by the API.
}
\examples{
\dontrun{
# Fetch annual GDP data for all industries in 2022
data <- beaGDPbyIndustry(TableID = "1", Industry = "ALL", Frequency = "A", Year = "2022")
}

}
\seealso{
\url{https://apps.bea.gov/API/signup/} for BEA API registration
}
