% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaInOut.R
\name{beaInOut}
\alias{beaInOut}
\title{Retrieve Input-Output Data from the Bureau of Economic Analysis (BEA) API}
\usage{
beaInOut(TableID = "", Year = "", ResultFormat = "json", beaKey = NULL)
}
\arguments{
\item{TableID}{A character string specifying the ID of the table to retrieve.}

\item{Year}{A character string specifying the year for which to retrieve data.}

\item{ResultFormat}{A character string specifying the format of the results. Default is "json".}

\item{beaKey}{An optional character string containing the BEA API key. If NULL, the function will attempt to retrieve the key using `getbeaKey()`.}
}
\value{
A data frame containing the requested Input-Output data. If an error occurs, it returns a data frame with error information.
}
\description{
This function fetches Input-Output data from the BEA API for a specified table and year.
}
\details{
This function requires a valid BEA API key. If not provided, it attempts to retrieve one using getbeaKey().
The function will issue a warning if the API key is invalid or if the API request results in an error.
}
\examples{
\dontrun{
# Retrieve data for a specific table and year
data <- beaInOut(TableID = "59", Year = "2022")

# Use a specific API key
data <- beaInOut(TableID = "262", Year = "2022", beaKey = "YOUR_API_KEY")
}

}
\seealso{
\url{https://apps.bea.gov/API/signup/} for BEA API registration
\code{\link{getbeaKey}} for retrieving a stored BEA API key
}
