% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaNIPADetails.R
\name{beaNIPADetails}
\alias{beaNIPADetails}
\title{Retrieve NIPA (National Income and Product Accounts) Details from BEA API}
\usage{
beaNIPADetails(
  TableName = "",
  Frequency = "",
  Year = "",
  ResultFormat = "json",
  beaKey = NULL
)
}
\arguments{
\item{TableName}{A character string specifying the NIPA table name.}

\item{Frequency}{A character string indicating the frequency of the data (e.g., "A" for annual, "Q" for quarterly).}

\item{Year}{A character string or numeric value representing the year(s) for which data is requested.}

\item{ResultFormat}{A character string specifying the format of the result. Default is "json".}

\item{beaKey}{An optional character string containing the BEA API key. If NULL, the function will attempt to retrieve it using `getbeaKey()`.}
}
\value{
A dataframe containing the requested NIPA data.
}
\description{
This function fetches NIPA details from the Bureau of Economic Analysis (BEA) API
}
\details{
This function requires a valid BEA API key. If not provided, it attempts to retrieve one using `getbeaKey()`.
Users can register for an API key at https://apps.bea.gov/API/signup/
}
\examples{
\dontrun{
# Fetch annual data for a specific table and year
data <- beaNIPADetails(TableName = "T10101", Frequency = "A", Year = "2022")

# Fetch quarterly data for a specific table and multiple years
data <- beaNIPADetails(TableName = "T10101", Frequency = "Q", Year = "2021,2022")
}

}
