% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaParamValues.R
\name{beaParamValues}
\alias{beaParamValues}
\title{Get Parameter Values for BEA Dataset}
\usage{
beaParamValues(
  DatasetName = "",
  ParameterName = "",
  ResultFormat = "json",
  beaKey = NULL
)
}
\arguments{
\item{DatasetName}{character. The name of the BEA dataset to query.}

\item{ParameterName}{character. The name of the parameter for which to retrieve values.}

\item{ResultFormat}{character. The format of the API response. Currently only "json" is supported. Default is "json".}

\item{beaKey}{character. Optional. The BEA API key. If NULL (default), it will be retrieved using `getbeaKey()`.}
}
\value{
A data frame containing available values for the specified parameter in the given dataset.
        If an error occurs, it returns a data frame with error information.
}
\description{
Retrieves a data frame of available values for a specified parameter
within a given Bureau of Economic Analysis (BEA) dataset.
}
\details{
This function interacts with the BEA API to fetch parameter values for a specified dataset.
It requires a valid BEA API key, which can be set using the `setbeaKey()` function or
provided directly as an argument.
}
\examples{
# Get table names for the Regional dataset
beaParamValues(DatasetName = "Regional", ParameterName = "TableName")

# Get available FIPS codes for the Regional dataset
beaParamValues(DatasetName = "Regional", ParameterName = "GeoFips")

}
\seealso{
\code{\link{setbeaKey}} for setting the BEA API key
\code{\link{getbeaKey}} for retrieving the stored BEA API key
\code{\link{beaParamList}} for getting a list of available parameters for a dataset
}
