% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaRegional.R
\name{beaRegional}
\alias{beaRegional}
\title{Retrieve Regional BEA Tables}
\usage{
beaRegional(
  TableName = "",
  LineCode = "",
  GeoFips = "",
  Year = "",
  ResultFormat = "json",
  beaKey = NULL
)
}
\arguments{
\item{TableName}{character. Specific table value to request (e.g., "CAGDP9").}

\item{LineCode}{numeric or character. Specific line from the specified table to request.}

\item{GeoFips}{character. Geographic area(s) for the requested data. Use FIPS codes.}

\item{Year}{numeric or character. Year(s) for which data is requested.}

\item{ResultFormat}{character. Format of the returned data. Currently only "json" is supported.}

\item{beaKey}{character. BEA API key. If NULL, searches for key in system environment.}
}
\value{
A data frame containing the requested BEA data with the following columns:
  \itemize{
    \item GeoFips: Geographic FIPS code
    \item GeoName: Name of the geographic area
    \item Code: Line code
    \item TimePeriod: Year of the data
    \item CL_UNIT: Classification unit
    \item UNIT_MULT: Unit multiplier
    \item DataValue: The actual data value (numeric)
    \item [Statistic]: The name of the statistic (column name varies)
  }
}
\description{
This function fetches data from Regional BEA Tables. It allows users to specify
table names, line codes, geographic areas, and years for which they need data.
Table names and line codes can be discovered using the `beaParamValues` function.
}
\note{
- The function will display a message with the statistic name and print any notes returned by the API.
- If an error occurs, the function will return the error information as a data frame.
}
\examples{
\dontrun{
# Fetch GDP data for the entire United States in 2022
gdp_data <- beaRegional(
  TableName = "CAGDP9",
  LineCode = 11,
  GeoFips = "00000",
  Year = 2022
)

# Fetch data for multiple years
multi_year_data <- beaRegional(
  TableName = "CAGDP9",
  LineCode = 11,
  GeoFips = "00000",
  Year = "2020,2021,2022"
)
}

}
\seealso{
\code{\link{beaParamValues}} for discovering table names and line codes.
\code{\link{setbeaKey}} for setting the BEA API key.
}
