% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSO-package.R
\docType{package}
\name{OSO-package}
\alias{OSO}
\alias{OSO-package}
\title{OSO: BEA R Package}
\description{
Make Requests from the Bureau of Economic Analysis API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aberuiz.github.io/OSO/}
}

}
\keyword{internal}
