% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setbeaKey.R
\name{setbeaKey}
\alias{setbeaKey}
\title{Set a BEA API key}
\usage{
setbeaKey(APIkey, install = FALSE, overwrite = FALSE)
}
\arguments{
\item{APIkey}{36 character string given to you by bea.gov}
}
\value{
Your BEA API key is set for data requests
}
\description{
@description
 Every request requires a BEA API key to be set in order to complete a request.
}
