% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaParamValues.R
\name{beaParamValues}
\alias{beaParamValues}
\title{Return data frame of parameter values}
\usage{
beaParamValues(
  DatasetName = "",
  ParameterName = "",
  ResultFormat = "json",
  beaKey = NULL
)
}
\arguments{
\item{DatasetName}{The name of the dataset}

\item{ParameterName}{Provides values for the named parameter}

\item{ResultFormat}{Currently OSO can only return data properly in json}

\item{beaKey}{Searches Sys.getenv by default. You may provide your API key here or save one with `setbeaKey`}
}
\value{
data frame of available values for the provided parameter for provided dataset
}
\description{
Returns a data frame of available values for the provided parameter for provided dataset
}
\examples{
beaParamValues(DataSetName = "Regional", ParameterName = "TableName")

beaParamValues(DataSetName = "Regional", ParameterName = "GeoFips")
}
