% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaParamValuesFiltered.R
\name{beaParamValuesFiltered}
\alias{beaParamValuesFiltered}
\title{Return a data frame of filtered parameter values}
\usage{
beaParamValuesFiltered(
  DatasetName = "",
  TargetParameter = "",
  TableName = "",
  LineCode = "",
  ResultFormat = "json",
  beaKey = NULL
)
}
\arguments{
\item{DatasetName}{The name of the dataset}

\item{TableName}{Specific BEA table to filter}

\item{LineCode}{Specific line to filter}

\item{ResultFormat}{Currently OSO can only return data properly in json}

\item{beaKey}{Searches Sys.getenv by default. You may provide your API key here or save one with `setbeaKey`}

\item{ParameterName}{Provides values for the named parameter}
}
\value{
data frame of available values for the provided parameter for provided dataset
}
\description{
Returns a data frame of available parameter values as filtered by another parameter
}
\examples{
beaParamValuesFiltered(
  DatasetName = "Regional",
  TargetParameter = "linecode",
  TableName = "CAGDP9"
  )
}
