% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaRegional.R
\name{beaRegional}
\alias{beaRegional}
\title{Return Regional BEA Tables}
\usage{
beaRegional(
  TableName = "",
  LineCode = "",
  GeoFips = "",
  Year = "",
  ResultFormat = "json",
  beaKey = NULL
)
}
\arguments{
\item{TableName}{Specific table value to request}

\item{LineCode}{Specific line from the specified table to request}

\item{GeoFips}{Geographic area(s) for the above data}

\item{Year}{Year(s) requested}

\item{ResultFormat}{Currently OSO can only return data properly in json}

\item{beaKey}{Searches Sys.getenv by default. You may provide your API key here or save one with `setbeaKey`}
}
\value{
Table and Line code values from BEA in a data frame
}
\description{
Returns Regional BEA Tables. Table names and Line codes can be discovered with `beaParamValues`
}
\examples{
beaRegional(
  TableName = "CAGDP9",
  LineCode = 11,
  GeoFips = "00000",
  Year = 2022
)

}
